Discard task events during cleanup period
Relates to #313

This fixes a deadlock that could occur during the period of time between
when a container waited for cleanup and when the actual cleanup finished
(removes and state management changes).

During that period of time, the task_manager would not accept any new
events and, immediately after that period of time, the task_manager
needed to take the processTasks lock to complete removing itself.

Due to another issue, the scope of that lock when sending events was
overly broad, leading to a deadlock.

This fixes that deadlock by appropriately reducing the scope of the
processTasks lock when handling events.

It also speeds up the processing of multiple tasks by continuing to read
task events during the cleanup actions. Prior to this, while a task's
container's were being removed, no other task could process new events.